Moved focus into own module
