+++ src/layout/actions/mod.rs
@@ -1,2 +1,3 @@
 pub mod layout;
 pub mod movement;
+pub mod focus;
