+++ src/layout/commands.rs
@@ -249,7 +249,7 @@ impl Tree {
 
     /// Sets the view to be the new active container. Never fails
     pub fn set_active_view(&mut self, view: WlcView) -> CommandResult {
-        self.0.set_active_container(view.clone());
+        self.0.set_active_view(view.clone());
         view.focus();
         Ok(())
     }
