+++ src/layout/core/tree.rs
@@ -9,7 +9,7 @@ use rustwlc::{WlcView, WlcOutput, Geometry, Point};
 
 use super::super::LayoutTree;
 
-use super::container::{Container, Handle, ContainerType, Layout};
+use super::container::{Container, ContainerType, Layout};
 
 use super::super::core::graph_tree::GraphError;
 
@@ -56,10 +56,29 @@ impl LayoutTree {
     }
 
     /// Sets the active container by finding the node with the WlcView
-    pub fn set_active_container(&mut self, handle: WlcView) {
+    pub fn set_active_view(&mut self, handle: WlcView) {
         if let Some(node_ix) = self.tree.descendant_with_handle(self.tree.root_ix(), &handle) {
             self.set_active_node(node_ix);
         }
+    }
+
+    /// Sets the active container associated with the UUID to be the active container.
+    ///
+    /// If the container was not a view or container, or the UUID was invalid,
+    /// then an error is returned.
+    pub fn set_active_container(&mut self, id: Uuid) -> Result<(), TreeError>{
+        let node_ix = try!(self.tree.lookup_id(id)
+                           .ok_or(TreeError::NodeNotFound(id)));
+        match self.tree[node_ix].get_type() {
+            ContainerType::View { .. } | ContainerType::Container { .. } => {
+                self.active_container = Some(node_ix);
+                Ok(())
+            },
+            _ => {
+                Err(TreeError::UuidWrongType(id, vec!(ContainerType::View, ContainerType::Container)))
+            }
+        }
+
     }
 
     /// Sets the active container to be the given node.
@@ -379,119 +398,6 @@ impl LayoutTree {
             warn!("Cannot move active focus when not there is no active container");
         }
         self.validate();
-    }
-    fn move_focus_recurse(&mut self, node_ix: NodeIndex, direction: Direction) -> Result<NodeIndex, ()> {
-        match self.tree[node_ix].get_type() {
-            ContainerType::View | ContainerType::Container => { /* continue */ },
-            _ => return Err(())
-        }
-        let parent_ix = self.tree.parent_of(node_ix)
-            .expect("Active ix had no parent");
-        match self.tree[parent_ix] {
-            Container::Container { layout, .. } => {
-                match (layout, direction) {
-                    (Layout::Horizontal, Direction::Left) |
-                    (Layout::Horizontal, Direction::Right) |
-                    (Layout::Vertical, Direction::Up) |
-                    (Layout::Vertical, Direction::Down) => {
-                        let siblings = self.tree.children_of(parent_ix);
-                        let cur_index = siblings.iter().position(|node| {
-                            *node == node_ix
-                        }).expect("Could not find self in parent");
-                        let maybe_new_index = match direction {
-                            Direction::Right | Direction::Down => {
-                                cur_index.checked_add(1)
-                            }
-                            Direction::Left  | Direction::Up => {
-                                cur_index.checked_sub(1)
-                            }
-                        };
-                        if maybe_new_index.is_some() &&
-                            maybe_new_index.unwrap() < siblings.len() {
-                                // There is a sibling to move to.
-                                let new_index = maybe_new_index.unwrap();
-                                let new_active_ix = siblings[new_index];
-                                match self.tree[new_active_ix].get_type() {
-                                    ContainerType::Container => {
-                                        // Get the first view we can find in the container
-                                        let first_view = self.tree.descendant_of_type(new_active_ix, ContainerType::View)
-                                            .expect("Could not find view in ancestor sibling container");
-                                        trace!("Moving to different view {:?} in container {:?}",
-                                               self.tree[first_view], self.tree[new_active_ix]);
-                                        return Ok(first_view);
-                                    },
-                                    ContainerType::View => {
-                                        trace!("Moving to other view {:?}", self.tree[new_active_ix]);
-                                        return Ok(new_active_ix)
-                                    },
-                                    _ => unreachable!()
-                                };
-                            }
-                    },
-                    _ => { /* We are moving out of siblings, recurse */ }
-                }
-            }
-            Container::Workspace { .. } => {
-                return Err(());
-            }
-            _ => unreachable!()
-        }
-        let parent_ix = self.tree.parent_of(node_ix)
-            .expect("Node had no parent");
-        return self.move_focus_recurse(parent_ix, direction);
-    }
-
-    /// Updates the current active container to be the next container or view
-    /// to focus on after the previous view/container was moved/removed.
-    ///
-    /// A new view will tried to be set, starting with the children of the
-    /// parent node. If a view cannot be found there, it starts climbing the
-    /// tree until either a view is found or the workspace is (in which case
-    /// it set the active container to the root container of the workspace)
-    fn focus_on_next_container(&mut self, mut parent_ix: NodeIndex) {
-        while self.tree.node_type(parent_ix)
-            .expect("Node not part of the tree") != ContainerType::Workspace {
-            if let Some(view_ix) = self.tree.descendant_of_type_right(parent_ix,
-                                                           ContainerType::View) {
-                match self.tree[view_ix]
-                                    .get_handle().expect("view had no handle") {
-                    Handle::View(view) => view.focus(),
-                    _ => panic!("View had an output handle")
-                }
-                trace!("Active container set to view at {:?}", view_ix);
-                self.set_active_node(view_ix);
-                return;
-            }
-            parent_ix = self.tree.ancestor_of_type(parent_ix,
-                                                   ContainerType::Container)
-                .unwrap_or_else(|| {
-                    self.tree.ancestor_of_type(parent_ix, ContainerType::Workspace)
-                        .expect("Container was not part of a workspace")
-                });
-        }
-        // If this is reached, parent is workspace
-        let container_ix = self.tree.children_of(parent_ix)[0];
-        let root_c_children = self.tree.children_of(container_ix);
-        if root_c_children.len() > 0 {
-            let new_active_ix = self.tree.descendant_of_type(root_c_children[0],
-                                                             ContainerType::View)
-                .unwrap_or(root_c_children[0]);
-            self.set_active_node(new_active_ix);
-            match self.tree[new_active_ix] {
-                Container::View { ref handle, .. } => handle.focus(),
-                _ => {}
-            };
-            return;
-        }
-        trace!("Active container set to container {:?}", container_ix);
-        self.set_active_node(container_ix);
-
-        // Update focus to new container
-        self.get_active_container().map(|con| match *con {
-            Container::View { ref handle, .. } => handle.focus(),
-            Container::Container { .. } => WlcView::root().focus(),
-            _ => panic!("Active container not view or container!")
-        });
     }
 
     /// Changes the layout of the active container to the given layout.
@@ -924,7 +830,7 @@ pub mod tests {
          */
         assert!(tree.active_ix_of(ContainerType::Output).is_some());
         assert!(tree.active_ix_of(ContainerType::Root).is_some());
-        tree.set_active_container(WlcView::root());
+        tree.set_active_view(WlcView::root());
         let view_ix = tree.tree.descendant_with_handle(tree.tree.root_ix(), &WlcView::root()).unwrap();
         assert_eq!(tree.active_container, Some(view_ix));
         tree.unset_active_container();
